% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{identity function whose derivative is exp(x)}
\usage{
myfun(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
identity function whose derivative is exp(x)
}
\examples{
Deriv(myfun)
}
